+++ CHANGELOG.md
@@ -19,6 +19,7 @@ x.y.z Release notes (yyyy-MM-dd)
 * Add the following languages: Czech, French, Portuguese, Dutch, Italian, Polish, Romanian, Russian and Spanish
 * Regenerate all API files with a new headers parameters per request
 * Add process payment quick fix for pre auth payments
+* Add new property `isOnlineCoFuGasStation` to `GasStation`
 
 
 13.0.0 Release notes (2022-02-10)